// src/components/types.ts
export type Company = {
  allocation: number;
  name: string;
  ticker: string;
  currency: string;
  sector: string;
  countryCode: string;
  continentCode: string;
  dividendYieldTTM: number;
  stockPrice: number;
  marketCapSize: string;
  shareInChosenCurrency: number;
  shareInStockCurrency: number;
  nbStocks: number;
};